Refactor datatables into separate files
